code cleanup: remove paranoid/invalid NULL checks and also reduce some unneeded size_t -> int conversions.

+++ source/blender/blenfont/intern/blf_font.c
@@ -159,7 +159,7 @@ static void blf_font_ensure_ascii_table(FontBLF *font)
 	}                                                                            \
 } (void)0
 
-void blf_font_draw(FontBLF *font, const char *str, unsigned int len)
+void blf_font_draw(FontBLF *font, const char *str, size_t len)
 {
 	unsigned int c;
 	GlyphBLF *g, *g_prev = NULL;
@@ -191,7 +191,7 @@ void blf_font_draw(FontBLF *font, const char *str, unsigned int len)
 }
 
 /* faster version of blf_font_draw, ascii only for view dimensions */
-void blf_font_draw_ascii(FontBLF *font, const char *str, unsigned int len)
+void blf_font_draw_ascii(FontBLF *font, const char *str, size_t len)
 {
 	unsigned char c;
 	GlyphBLF *g, *g_prev = NULL;
